requirements.txt
python-decouple==3.8
PyYAML==6.0
pymongo==4.6.2
requests==2.31.0
pandas==1.5.3
openpyxl==3.0.10